pyproject.toml
[project]
name = "njjc-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "asgiref==3.11.0",
    "attrs==25.4.0",
    "boto3==1.42.14",
    "botocore==1.42.14",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "click==8.3.1",
    "cryptography==46.0.3",
    "django==6.0",
    "django-cors-headers==4.9.0",
    "django-storages==1.14.6",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "jmespath==1.0.1",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "markdown-it-py==4.0.0",
    "mcp[cli]==1.24.0",
    "mdurl==0.1.2",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "pillow==12.0.0",
    "psycopg2-binary==2.9.11",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "referencing==0.37.0",
    "requests>=2.32.5",
    "rich==14.2.0",
    "rpds-py==0.30.0",
    "s3transfer==0.16.0",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sqlparse==0.5.4",
    "sse-starlette==3.0.4",
    "starlette==0.50.0",
    "transformers>=4.57.3",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "uvicorn==0.38.0",
]
